requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-api-key==2.0.0
djangorestframework-camel-case==1.2.0
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
